Migrate to MAUI framework
