+++ ClipCopy.sln
@@ -1,8 +1,8 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ClipCopy", "ClipCopy\ClipCopy.csproj", "{6AA0EF8D-31D7-4A71-AAB1-75204B1007AC}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ClipCopy", "ClipCopy\ClipCopy.csproj", "{D363D663-3D88-4373-9A26-3824B8121D23}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ClipCopy.Database", "ClipCopy.Database\ClipCopy.Database.csproj", "{9C483DF3-C345-4247-9346-22E5A0CD6A59}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ClipCopy.Database", "ClipCopy.Database\ClipCopy.Database.csproj", "{8CCCBE14-3F6C-4C2C-9872-2717116525CB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -10,15 +10,13 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{6AA0EF8D-31D7-4A71-AAB1-75204B1007AC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6AA0EF8D-31D7-4A71-AAB1-75204B1007AC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6AA0EF8D-31D7-4A71-AAB1-75204B1007AC}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
-		{6AA0EF8D-31D7-4A71-AAB1-75204B1007AC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6AA0EF8D-31D7-4A71-AAB1-75204B1007AC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6AA0EF8D-31D7-4A71-AAB1-75204B1007AC}.Release|Any CPU.Deploy.0 = Release|Any CPU
-		{9C483DF3-C345-4247-9346-22E5A0CD6A59}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9C483DF3-C345-4247-9346-22E5A0CD6A59}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9C483DF3-C345-4247-9346-22E5A0CD6A59}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9C483DF3-C345-4247-9346-22E5A0CD6A59}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D363D663-3D88-4373-9A26-3824B8121D23}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D363D663-3D88-4373-9A26-3824B8121D23}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D363D663-3D88-4373-9A26-3824B8121D23}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D363D663-3D88-4373-9A26-3824B8121D23}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8CCCBE14-3F6C-4C2C-9872-2717116525CB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8CCCBE14-3F6C-4C2C-9872-2717116525CB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8CCCBE14-3F6C-4C2C-9872-2717116525CB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8CCCBE14-3F6C-4C2C-9872-2717116525CB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
