+++ ClipCopy/Platforms/Android/Activities/MainActivity.cs
@@ -1,0 +1,11 @@
+﻿using Android.App;
+using Android.Content.PM;
+
+namespace ClipCopy.Platforms.Android.Activities;
+
+[Activity(Theme = "@style/Maui.SplashTheme", MainLauncher = true,
+    ConfigurationChanges = ConfigChanges.ScreenSize | ConfigChanges.Orientation | ConfigChanges.UiMode |
+                           ConfigChanges.ScreenLayout | ConfigChanges.SmallestScreenSize | ConfigChanges.Density)]
+public class MainActivity : MauiAppCompatActivity
+{
+}
