+++ ClipCopy/Platforms/Android/Resources/values-pl/strings.xml
