+++ ClipCopy/Platforms/Android/Resources/values/strings.xml
