+++ ClipCopy/Platforms/Android/Resources/values/styles.xml
