+++ ClipCopy/Program.cs
@@ -1,0 +1,18 @@
+﻿using Microsoft.Extensions.Logging;
+
+namespace ClipCopy;
+
+public static class Program
+{
+    public static MauiApp CreateMauiApp()
+    {
+        var builder = MauiApp.CreateBuilder();
+        builder.UseMauiApp<App>();
+
+#if DEBUG
+        builder.Logging.AddDebug();
+#endif
+
+        return builder.Build();
+    }
+}
