--- ClipCopy/Resources/Localization.pl-PL.resx
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<root>
-    <xsd:schema id="root" xmlns="" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns:msdata="urn:schemas-microsoft-com:xml-msdata">
-        <xsd:element name="root" msdata:IsDataSet="true">
-
-        </xsd:element>
-    </xsd:schema>
-    <resheader name="resmimetype">
-        <value>text/microsoft-resx</value>
-    </resheader>
-    <resheader name="version">
-        <value>1.3</value>
-    </resheader>
-    <resheader name="reader">
-        <value>System.Resources.ResXResourceReader, System.Windows.Forms, Version=2.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
-    </resheader>
-    <resheader name="writer">
-        <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=2.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
-    </resheader>
-    <data name="clipboard_copied" xml:space="preserve">
-        <value>Skopiowano {0}!</value>
-    </data>
-    <data name="clipboard_send_to" xml:space="preserve">
-        <value>Wyślij do schowka</value>
-    </data>
-    <data name="clipboard_service_failed" xml:space="preserve">
-        <value>Błąd pobierania serwisu schowka</value>
-    </data>
-</root>
