+++ ClipCopy/Resources/Localizations/Localization.pl-pl.resx
@@ -1,0 +1,23 @@
+<root>
+    <resheader name="resmimetype">
+        <value>text/microsoft-resx</value>
+    </resheader>
+    <resheader name="version">
+        <value>1.3</value>
+    </resheader>
+    <resheader name="reader">
+        <value>System.Resources.ResXResourceReader, System.Windows.Forms, Version=2.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
+    </resheader>
+    <resheader name="writer">
+        <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=2.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
+    </resheader>
+    <data name="clipboard_copied" xml:space="preserve">
+        <value>Skopiowano {0}!</value>
+    </data>
+    <data name="clipboard_service_failed" xml:space="preserve">
+        <value>Błąd pobierania serwisu schowka!</value>
+    </data>
+    <data name="clipboard_send_to" xml:space="preserve">
+        <value>Wyślij do schowka</value>
+    </data>
+</root>
