--- ClipCopy/Resources/mipmap-anydpi-v26/appicon.xml
@@ -1,4 +1,0 @@
-<adaptive-icon xmlns:android="http://schemas.android.com/apk/res/android">
-  <background android:drawable="@mipmap/appicon_background" />
-  <foreground android:drawable="@mipmap/appicon_foreground" />
-</adaptive-icon>
