+++ ClipCopy/Views/App.xaml.cs
@@ -1,0 +1,11 @@
+﻿namespace ClipCopy;
+
+public partial class App
+{
+    public App()
+    {
+        InitializeComponent();
+
+        MainPage = new AppShell();
+    }
+}
