+++ ClipCopy/Views/AppShell.xaml.cs
@@ -1,0 +1,9 @@
+﻿namespace ClipCopy;
+
+public partial class AppShell
+{
+    public AppShell()
+    {
+        InitializeComponent();
+    }
+}
