+++ ClipCopy/Views/MainPage.xaml.cs
@@ -1,0 +1,9 @@
+﻿namespace ClipCopy.Views;
+
+public partial class MainPage
+{
+    public MainPage()
+    {
+        InitializeComponent();
+    }
+}
